Fix \x escapes truncating above 0x00FF
Both escape switches accumulate the four hex digits of \xXXXX into a `byte`, so
the third and fourth `b <<= 4` discard the high bits: \x20AC came out as 0x00AC,
\x0100 as 0, \xFFFF as 0x00FF. \u was never affected - it accumulates into a
`ushort`.

Same failure mode as the cursor bug this branch already fixes, and in the same
two methods: nothing reports an error, the generator exits 0 and the table
compiles, the grammar just means a different character. Widening `b` to `int` at
all four sites fixes it.

This is a pre-existing defect. `byte b=_FromHexChar` is on master at all four
sites and the cursor fix only added `pc.Advance()`, never touching the
accumulator width.

No golden moves. No grammar in the repository uses \x with more than two hex
digits - the only \x in Slang.rl is the literal text `\x[0-9A-Fa-f]{2}`, which
describes a two-digit escape rather than being one - so all eight hashes in
testcases/expected.txt and all 39 rows of fa-golden.tsv are unchanged.

EscapeParsingTests gains eight cases covering both front ends, ordinary and
charset, plus the sub-0x100 values that must not move. They fail on the unfixed
accumulator and pass after.

+++ Rolex/FA.brick.cs
@@ -1146,14 +1146,14 @@ expr=Repeat(expr,min,max,accept,compact);break;}return expr;}static byte _FromHe
 <hex)return(byte)(hex-'7'); if('g'>hex&&'`'<hex)return(byte)(hex-'W'); throw new ArgumentException("The value was not hex.","hex");}static bool _IsHexChar(int
  hex){if(':'>hex&&'/'<hex)return true;if('G'>hex&&'@'<hex)return true;if('g'>hex&&'`'<hex)return true;return false;} static int _ParseEscapePart(LexContext
  pc){if(-1==pc.Current)return-1;switch(pc.Current){case'f':pc.Advance();return'\f';case'v':pc.Advance();return'\v';case't':pc.Advance();return'\t';case
-'n':pc.Advance();return'\n';case'r':pc.Advance();return'\r';case'x':if(-1==pc.Advance()||!_IsHexChar(pc.Current))return'x';byte b=_FromHexChar(pc.Current);
+'n':pc.Advance();return'\n';case'r':pc.Advance();return'\r';case'x':if(-1==pc.Advance()||!_IsHexChar(pc.Current))return'x';int b=_FromHexChar(pc.Current);
 if(-1==pc.Advance()||!_IsHexChar(pc.Current))return unchecked(b);b<<=4;b|=_FromHexChar(pc.Current);if(-1==pc.Advance()||!_IsHexChar(pc.Current))return
  unchecked(b);b<<=4;b|=_FromHexChar(pc.Current);if(-1==pc.Advance()||!_IsHexChar(pc.Current))return unchecked(b);b<<=4;b|=_FromHexChar(pc.Current);pc.Advance();return
  unchecked(b);case'u':if(-1==pc.Advance())return'u';ushort u=_FromHexChar(pc.Current);u<<=4;if(-1==pc.Advance())return unchecked(u);u|=_FromHexChar(pc.Current);
 u<<=4;if(-1==pc.Advance())return unchecked(u);u|=_FromHexChar(pc.Current);u<<=4;if(-1==pc.Advance())return unchecked(u);u|=_FromHexChar(pc.Current);pc.Advance();return
  unchecked(u);default:int i=pc.Current;pc.Advance();return i;}}static int _ParseRangeEscapePart(LexContext pc){if(-1==pc.Current)return-1;switch(pc.Current)
 {case'0':pc.Advance();return'\0';case'f':pc.Advance();return'\f';case'v':pc.Advance();return'\v';case't':pc.Advance();return'\t';case'n':pc.Advance();
-return'\n';case'r':pc.Advance();return'\r';case'x':if(-1==pc.Advance()||!_IsHexChar(pc.Current))return'x';byte b=_FromHexChar(pc.Current);if(-1==pc.Advance()
+return'\n';case'r':pc.Advance();return'\r';case'x':if(-1==pc.Advance()||!_IsHexChar(pc.Current))return'x';int b=_FromHexChar(pc.Current);if(-1==pc.Advance()
 ||!_IsHexChar(pc.Current))return unchecked(b);b<<=4;b|=_FromHexChar(pc.Current);if(-1==pc.Advance()||!_IsHexChar(pc.Current))return unchecked(b);b<<=4;
 b|=_FromHexChar(pc.Current);if(-1==pc.Advance()||!_IsHexChar(pc.Current))return unchecked(b);b<<=4;b|=_FromHexChar(pc.Current);pc.Advance();return unchecked(b);case
 'u':if(-1==pc.Advance())return'u';ushort u=_FromHexChar(pc.Current);u<<=4;if(-1==pc.Advance())return unchecked(u);u|=_FromHexChar(pc.Current);u<<=4;if
@@ -1702,13 +1702,13 @@ internal static string EscapeRangeChar(string character){var codepoint=char.Conv
 "hex");}static bool _IsHexChar(int hex){if(':'>hex&&'/'<hex)return true;if('G'>hex&&'@'<hex)return true;if('g'>hex&&'`'<hex)return true;return false;}
  static int _ParseEscapePart(LexContext pc){if(-1==pc.Current)return-1;switch(pc.Current){case'f':pc.Advance();return'\f';case'v':pc.Advance();return'\v';
 case't':pc.Advance();return'\t';case'n':pc.Advance();return'\n';case'r':pc.Advance();return'\r';case'x':if(-1==pc.Advance()||!_IsHexChar(pc.Current))return
-'x';byte b=_FromHexChar(pc.Current);if(-1==pc.Advance()||!_IsHexChar(pc.Current))return unchecked(b);b<<=4;b|=_FromHexChar(pc.Current);if(-1==pc.Advance()
+'x';int b=_FromHexChar(pc.Current);if(-1==pc.Advance()||!_IsHexChar(pc.Current))return unchecked(b);b<<=4;b|=_FromHexChar(pc.Current);if(-1==pc.Advance()
 ||!_IsHexChar(pc.Current))return unchecked(b);b<<=4;b|=_FromHexChar(pc.Current);if(-1==pc.Advance()||!_IsHexChar(pc.Current))return unchecked(b);b<<=4;
 b|=_FromHexChar(pc.Current);pc.Advance();return unchecked(b);case'u':if(-1==pc.Advance())return'u';ushort u=_FromHexChar(pc.Current);u<<=4;if(-1==pc.Advance())return
  unchecked(u);u|=_FromHexChar(pc.Current);u<<=4;if(-1==pc.Advance())return unchecked(u);u|=_FromHexChar(pc.Current);u<<=4;if(-1==pc.Advance())return unchecked(u);
 u|=_FromHexChar(pc.Current);pc.Advance();return unchecked(u);default:int i=pc.Current;pc.Advance();return i;}}static int _ParseRangeEscapePart(LexContext pc){if(-1==
 pc.Current)return-1;switch(pc.Current){case'f':pc.Advance();return'\f';case'v':pc.Advance();return'\v';case't':pc.Advance();return'\t';case'n':pc.Advance();
-return'\n';case'r':pc.Advance();return'\r';case'x':if(-1==pc.Advance()||!_IsHexChar(pc.Current))return'x';byte b=_FromHexChar(pc.Current);if(-1==pc.Advance()
+return'\n';case'r':pc.Advance();return'\r';case'x':if(-1==pc.Advance()||!_IsHexChar(pc.Current))return'x';int b=_FromHexChar(pc.Current);if(-1==pc.Advance()
 ||!_IsHexChar(pc.Current))return unchecked(b);b<<=4;b|=_FromHexChar(pc.Current);if(-1==pc.Advance()||!_IsHexChar(pc.Current))return unchecked(b);b<<=4;
 b|=_FromHexChar(pc.Current);if(-1==pc.Advance()||!_IsHexChar(pc.Current))return unchecked(b);b<<=4;b|=_FromHexChar(pc.Current);pc.Advance();return unchecked(b);case
 'u':if(-1==pc.Advance())return'u';ushort u=_FromHexChar(pc.Current);u<<=4;if(-1==pc.Advance())return unchecked(u);u|=_FromHexChar(pc.Current);u<<=4;if
